backend/chds_food_delivery/requirements.txt
django==5.1.3
python-dotenv==1.0.1
psycopg2==2.9.10
pillow==11.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3